requirements.txt
bs4
lxml
pyyaml
requests
tiktoken
openai